Proposers handle future promises
If a proposer receives a promise for a proposal whose id is greater than the
current proposal, (but whose proposer id is equal) then this indicates that the
proposal has been abandoned for the later one. Therefore discard all the
promises made so far and start again with the new proposal.